requirements.txt
discord
d20
python-dotenv
sqlite3